@@ -7,7 +7,7 @@ actual class ParcelableStub actual constructor() : Parcelable {
 
     override fun describeContents(): Int = 0
 
-    override fun writeToParcel(dest: Parcel?, flags: Int) {
+    override fun writeToParcel(dest: Parcel, flags: Int) {
     }
 
     companion object CREATOR : android.os.Parcelable.Creator<ParcelableStub> {
