@@ -1,6 +1,6 @@
 package com.arkivanov.decompose
 
-import com.arkivanov.essenty.backpressed.BackPressedHandlerOwner
+import com.arkivanov.essenty.backhandler.BackHandlerOwner
 import com.arkivanov.essenty.instancekeeper.InstanceKeeperOwner
 import com.arkivanov.essenty.lifecycle.LifecycleOwner
 import com.arkivanov.essenty.statekeeper.StateKeeperOwner
@@ -9,4 +9,4 @@ interface ComponentContext :
     LifecycleOwner,
     StateKeeperOwner,
     InstanceKeeperOwner,
-    BackPressedHandlerOwner
+    BackHandlerOwner
