@@ -9,11 +9,6 @@ import com.arkivanov.decompose.ExperimentalDecomposeApi
 enum class Direction {
 
     /**
-     * The child is not being animated.
-     */
-    IDLE,
-
-    /**
      * The child is entering from the front (push).
      */
     ENTER_FRONT,
@@ -33,3 +28,31 @@ enum class Direction {
      */
     EXIT_BACK,
 }
+
+@ExperimentalDecomposeApi
+val Direction.isEnter: Boolean
+    get() =
+        when (this) {
+            Direction.ENTER_FRONT,
+            Direction.ENTER_BACK -> true
+            Direction.EXIT_FRONT,
+            Direction.EXIT_BACK -> false
+        }
+
+@ExperimentalDecomposeApi
+val Direction.isExit: Boolean
+    get() = !isEnter
+
+@ExperimentalDecomposeApi
+val Direction.isFront: Boolean
+    get() =
+        when (this) {
+            Direction.ENTER_FRONT,
+            Direction.EXIT_FRONT -> true
+            Direction.ENTER_BACK,
+            Direction.EXIT_BACK -> false
+        }
+
+@ExperimentalDecomposeApi
+val Direction.isBack: Boolean
+    get() = !isFront
